+++ pyqcrbox/pyqcrbox/services/services_registry.py
@@ -7,22 +7,19 @@ from pyqcrbox.logging import logger
 
 
 def create_nats_broker():
-    return NatsBroker(settings.nats.url, graceful_timeout=10, max_reconnect_attempts=1, logger=logger)
+    return NatsBroker(
+        settings.nats.url,
+        graceful_timeout=settings.nats.graceful_timeout,
+        max_reconnect_attempts=settings.nats.max_reconnect_attempts,
+        logger=logger,
+    )
 
-class DevelopmentServicesRegistry(svcs.Registry):
-    def __init__(self):
-        super().__init__()
-
-        # self.register_value(DataFileManager, DummyDataFileManager())
-        self.register_value(DataFileManager, NatsDataFileManager())
-        self.register_factory(NatsBroker, create_nats_broker)
-
-
-def get_qcrbox_services_registry():
-    # For now, we're just using the development registry. In the future,
-    # we may want to use different registries for different environments
-    # (e.g., development, testing, production).
-    return DevelopmentServicesRegistry()
+async def create_data_file_manager(container: svcs.Container):
+    nats_broker = await container.aget(NatsBroker)
+    return NatsDataFileManager(nats_broker)
 
 
-QCRBOX_GLOBAL_SERVICES_REGISTRY = get_qcrbox_services_registry()
+
+QCRBOX_GLOBAL_SERVICES_REGISTRY = svcs.Registry()
+QCRBOX_GLOBAL_SERVICES_REGISTRY.register_factory(NatsBroker, create_nats_broker)
+QCRBOX_GLOBAL_SERVICES_REGISTRY.register_factory(DataFileManager, create_data_file_manager)